requirements.txt
gTTS==2.5.1
pygame==2.5.2

SpeechRecognition==3.10.4
sounddevice==0.4.7
wavio==0.0.7
numpy>=1.24
azure-cognitiveservices-speech==1.46.0
requests>=2.31.0

# Offline (opcional para STT):
# vosk==0.3.45
